Revert "Revert "working mamba for training""
This reverts commit 83d7cd1750fea75ca61e2c6754f778efa07dc70b.

+++ hswvime_ppo/buffer.py
@@ -112,7 +112,7 @@ class TransitionRolloutBuffer(RolloutBuffer):
         )
         self.questions = np.zeros(
             (self.buffer_size, self.n_envs, self.num_qa),
-            dtype=np.long,
+            dtype=np.float64,
         )
         self.answers = np.zeros(
             (self.buffer_size, self.n_envs, self.num_qa, self.answer_dim),
